Added Admin Configuration action, added the new configuration attributes to the module plus some tests and refactoring.
created the admin configuration view, added new methods to the ConfigFileService to retrieve the data.

Removed a redundant enum, changed the default values of ENABLED and DISABLED in the GeneralSettings enum to avoid issues with type during validation.

Changed the behaviour of some methods in the BasicEnum and added a method to retrieve automatically all the labels at once.

Added the new parameters to the module.

Changed the install script slightly to take into account the added parameters.

Created the ConfigurationForm object and added the necessary tests to it.

Fixed the ConfigFileService tests to avoid making them dependent to my local settings.