Fixed the installation, the EnumFilesService, some tests and the default Roles
The roles now inherit each other as intended.
Made the default permission for the standard user protected as well.
Changed the installation script to make it create the enum files if they are missing and alert the user before overriding them otherwise.
Added the required method in the EnumFilesService to check whether an enum file exists or not.
The EnumFilesService now checks the parent class that the enum is extending to avoid recreating the parent constants inside the child.
The service can also generate empty enums.

Fixed the tests that where broken by all the above changes plus created the tests for the new functionalities.